backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
click==8.1.7
h11==0.14.0
idna==3.7
pydantic_core==2.18.4
sniffio==1.3.1
starlette==0.37.2
typing-inspect==0.9.0
typing_extensions==4.12.2 